Less assumptions in twnmp-join-community-info and friends
* Being a translator doesn't imply being able to translate in the
  sandbox, for instance if the language is not supported.
* Being able to translate doesn't imply having time to submit some
  test translations immediately (or in the few hours usually allowed
  by translatewiki.net bureacrats before rejection).
* Having time to doesn't imply being willing to submit some trial
  translation to be later destroyed.

By reducing the chances that users are tricked into believing that
just by feeling they are "translators" they'll get an account magically
created, we should hopefully reduce failure rate and frustration.

Also, the question "Not a translator" didn't make sense under the
header says "Become a translator": using two definitions of "translator"
in the same form is not nice. Same for undefined "rights".

Bug: 70612
Change-Id: Iec63a1f1662b4c22e598302c3ec89885bfdc81f9