Make main page appear better in mobile devices
The interface does not scale well with scale ratio 1 on Nexus 4
for example. The grid system we use defines minimum width for row
as 600 px. Using 0.5 to make the page fit and to trigger the
one column layout using CSS.

Also gave a bit more space for the links as opposed for the site
name.

Change-Id: I6970411252c8b3b0a17c9929d74f968dbc43bd75