Fix cancel button styling for dev-signup
Easiest way to fix it was to clean up the styles. I added general
padding for the login-widget, so that there is no need to play with
margins of the elements inside it to have the left edge consistent.

This also avoids the need to use the weird "eleven columns" trick
for additional padding. Some of the selectors are a bit complicated
as we are overriding the default margings given by the grid.

Change-Id: I9219bca06855646ffb999dc25c6b69c7c980565f