Add way to store and fetch location coordinates
For lack of translatewiki specific extension, using TwnMainPage as
the closest equivalent instead.

This stores the locations in the page properties. New code is in src
and attempts to use latest best practices.

The "getter" only provides coordinates, not any info about them. I think
this is fine as the maps are mostly for illustration purposes. If needed,
we can join with the page table to get the user names.

To fully replace the existing method that uses SMW, the getter needs an
ability to filter by a category.

Bug: T346863
Change-Id: I9e31279c9d58df580686bc6226aff7676c3a78ec