build: Add dependency on Translate extension
Without this, Special:TwnMainPage crashes as follows:

```
Uncaught Error: Class "MediaWiki\Extension\Translate\Services" not found
Backtrace:
/mediawiki/extensions/TwnMainPage/includes/specials/SpecialTwnMainPage.php(31)
/mediawiki/vendor/wikimedia/object-factory/src/ObjectFactory.php(240)
/mediawiki/vendor/wikimedia/object-factory/src/ObjectFactory.php(149)
```

The dependency should make it a bit more likely that someone will
find it ahead of time, and when not, to find it in a clearer and
more direct way rather than after browsing to the page, e.g. when
setting it up locally to write a patch.

Change-Id: Iadcfd128d58d7feaea6ababdd98765b8043859cb