Fail early if $wgTranslateUseSandbox is disabled
$wgTranslateUseSandbox is disabled by default, so when installing
the Translate extension dependency, Special:MainPage appears to work
until you submit the form, when it fails with a confusing "badaction"
error, indicating that API action=translatesandbox is not defined.

It takes a bit of digging to realize that this actually comes from
an extension that is already installed (Translate) but is within
a conditional block that requires a separate feature flag.

Note that the extension has no mediawiki.org Extension page, and no
README. That's fine, but then an automatic nudge like this is all
the more useful :)

Change-Id: Ie79920dc611027d5eac2f864a9b492acd674e2cc